+++ packages/vscode-ws-jsonrpc/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this npm module are documented in this file.
 
+## [3.0.0] - upcoming
+
+- **BREAKING CHANGE**: Do not reexport code of imported libraries (e.g. vscode-languageclient) #[459](https://github.com/TypeFox/monaco-languageclient/pull/459)
+  - Content of `vscode-jsonrpc` is no longer re-exported
+
 ## [2.0.2] - 2022-01-24
 
 - Gracefully handle JSON.parse errors #[455](https://github.com/TypeFox/monaco-languageclient/pull/455)